Add ecall hooks
Adds capability to register an ecall receiver. A frontend can thereby
implement its own ecalls for example to visualize data. When there is
no standard ecall defined an ecall receiver is called.

An ecall hook is registered with `registerEcallReceiver`. The
prototype of the registered hook function is `(Int) -> Boolean` where
the parameter is the ecall id (`a0`). It returns whether this ID was
served by the ecall hook. The frontend-specific driver can access the
simulator to access state (registers).