Add ecall hooks
Adds capability to register an ecall receiver. A frontend can thereby
implement its own ecalls for example to visualize data. When there is
no standard ecall defined an ecall receiver is called.

The ecall hook is registered with `registerEcallReceiver`. The
prototype of the registered hook function is `(String) -> (String)`
where the String is json. This json contains the id and the other
argument registers as `a0` to `a7`.

If the hook returns `a0` and `a1` in the returned json, the registers
are updated accordingly.